Start using the new machinespec definition for creating the instance on vsphere
The code now acts accordingly based on the availability of the new cloud-init datasource
"VMwareGuestInfo" inside the template used. This availability can be set by the users in
the spec using the flag "vsphereCloudInit" set to true. For usual upstream version of
ubuntu cloud image that has been used so far, this flag need not be set at all and the
controller will utilize the OVF properties available on the ubuntu cloud image to inject
the cloud-init data.
Note: The new static IP support is only available with the templates that have the vsphere
cloud-init datasource. DHCP based IPs will continue to work as they have in the past.

Going forward we prefer users creating a custom ubuntu cloud image that also has this new
cloud-init datasource pre-installed. As that will give them the ability to use static IPs
as well if they want.

See https://github.com/akutz/cloud-init-vmware-guestinfo for more info on the new cloud-init
datasource

* Adds support for multiple nics
* Adds support for static IP

Resolves #24
Resolves #28

Change-Id: Id2dee12b382d57b6591f840dadd77a1c287667c4