Overload the VMTemplate parameter for search purposes
The MachineSpec.VMTemplate currently is treated as name of
the VM to be searched. In certain cases the name of the VM
can be duplicated inside a VC, thus we need a more precise
mechanism to point to a VM to be used as a template. For
this purpose we are now allowing the VMTemplate to also
be a valid UUID that would then be matched with the VMs
InstanceUUID property which is guaranteed to be unique.

The modified logic detects if the passed string looks like
a valid UUID then it tried to find the VM by InstaceUUID
if not found, then it fall back on treating the string as
the VM name.

Resolves #109

Change-Id: Ia5e6ca600a71ca784b54336299818b6ec892b949