Make the install and configure part of the scripts match the intent
Currently the configure part of the init script is also installing
packages, which should not be the case. The install part of the script
should be installing anything needed during the configure part.
This split is further used when we want to use template that is
pre-loaded with all the binaries. In that case, we only want to run
the configure part which would bring up the services

Resolves #97
Resolves #121

(cherry picked from commit ba8c883c64cd1be62fd25c7072efe1f53cdd4598)
Change-Id: I720eaf54829dd3f756dbfedb7691c004c54fcd01