Remove redundancies in the kustomize files
Also change the prefix to generate the provider specific names
This is important as the current scheme of naming would clash
with the objects created for the core cluster-api deployment

This change will also make it possible to deploy both the core
cluster-api controllers and the vsphere provider controllers in
the same namespace if needed. By default these are deployed in
different namespaces.

Resolves #168

Change-Id: Ie6d82393129549bacd505060b43b5cfa7c8de1ef