Use `go env` to get GOPATH (#171)
At least on my development machine I usually leave `$GOPATH` undefined and let Go use the default. This unfortunately then breaks a lot of scripts that assume `$GOPATH` is always set. I think using `go env GOPATH` is supposed to be the right way to get the current value, since it will give the default path if the environment variable is unset.