Update the self-service quickstart docs for minikube and kind
A recent update of the vendor code updated the clusterctl bootstrap support to include
kind.  This update includes information on using kind as a bootstrapper.

In addition, the minikube project has merged the Fusion support PR.  There is no
longer a need to jump through hoops to build a version of minikube that supports
Vmware Fusion.  This update simplifies the minikube installation section.

Resolves #189