Build CAPV images for Ubuntu/CentOS (#267)
* Build CAPV images for Ubuntu/CentOS

This patch introduces support for building CAPV images based on the
process originated with the CAPA project. Images are based on Ubuntu
18.04.2 LTS and CentOS 7.

Images also include the Cloud-Init Datasource for VMware GuestInfo.

* Updated the docs on how to test the images

This patch updates the documentation on how to test the images built
from `build/images`.

* Updated test cloud-init data with network v2

This patch updates the test cloud-init data to use a version 2 network
configuration.

* Update image docs for accessing remote images

This patch updates the image documentation with information on how to
access remotely deploying images for testing.

* Clarify the included cloudinit directory

This patch adds documentation to clarify that the files in the
`build/images/cloudinit` directory are for testing and not for use with
production systems.