Update the CRDs
This patch updates the CRDs with the updated version of the
controller-tools. The only change is that the additional "type: object"
lines were no longer required as both the Cluster and Machine provider
config objects inherited from ObjectMeta, itself a "type: object".

This also matches the layout of the CAPA CRDs.