Shell with GOPATH
This patch updates the Makefile to use a custom shell based on
"hack/shell-with-gopath.sh" to ensure that even when building outside
the GOPATH, K8s tooling will still work.

This patch also:

* Updates any go:generate code to use scripts in "hack/" since
go:generate ignores the current, working directory
* Removes the Makefile from "config/crds" and puts that logic into
"hack/update-generated.sh" since recursive Make does not use the custom
shell
* Remove the CRD validation added in
50dc7f0264761df87288f5aeea883c52401521f1. This will be replaced with a
new Prow pre-submit test that will call "hack/verify-crds.sh"

This patch fixes the one caveat with the approach of a nested GOPATH,
which was that if it was created in or out of a container, either the
host or the container wouldn't work the next time "make" was invoked.
This corrects that by checking first to see if the symlink is broken and
recreating it if it is.