Enhanced networking support
This patch adds support for enhanced network configuration:

* Multiple devices
* Multiple address per device
* Multiple address families per device
* Multiple gateways of multiple address families
* Global, static routes
* Per-device, static routes

To achieve this required refactoring the reconciliation logic in order
to separate the creation of VMs from when they are booted. All of this
is tied to the new Exists workflow, documented in the docs folder as a
diagram. This is key to understanding how CAPV works.