Update YAML generation to use Python
This patch updates the way the YAML manifests are generated to no longer
rely on envsubst and instead use Python, a language installed by default
on macOS and nearly all Linux distros. The changes include:

* The generation script is now located at hack/generate-yaml.sh
* The file hack/envsubst.py handles what used to require the gettext
package
* The image hack/tools/kustomize/Dockerfile is a Google distroless image
with the Kustomize tool. This image is only 20MIB in size.
* The generation script defaults to using the Kustomize image over a
local binary, but "-k local" can be used to prefer the local binary if
desired
* The manifest templates have been updated to account for API type
changes