Create simple make target for container images
The previous make targets prod-images, prod-push, ci-images, and ci-push
essentially did the same thing for prod vs CI. The prod targets were
never meant to be invoked by CI and used a hard-coded image version that
needed to be updated whenever a new tag was made. The CI targets are
invoked on every PR by CI, and pushed a new CAPV and clusterctl image
tagged with the git SHA.

Instead, we now combine them into a single set of targets, build-images
and push-images, that set the image version to a simpler version of `git
describe` that will always show the changes since the last annotated
tag. This conveys more meaning than a simple SHA, and the git hash is
still included. With that done, we can then just always build and push
images in the same fashion, with no distinction between CI and
Production. This will allow the CI system to build images for us both as
part of CI/testing, and post merge to master.