Modify -ci image makefile to default to build only
When running `make` with no arguments in scripts/e2e/bootstrap_job, the
default behavior was to build the image, push the image to gcr.io, and
delete the local copy of the image. That's not a useful default.
Instead, change the "all" target to just have "build" as a pre-req,
making "make" the equivalent of "make build". If someone wants to push
images, they can run "make push" instead. Deleting the local image is an
explicit step left to "make clean".