Add new flags to generate-yaml
This patch adds new flags to the generate-yaml script.

* -c can be used to specify the name of the cluster
* -m is used to specify the name of the CAPV manager image

If -c is specified then the output directory becomes
"./out/CLUSTER_NAME". If -o OUT_DIR is specified then the output
directory is "OUT_DIR". Use -o to specify explicit output directories,
otherwise the directories are unique based on the name of the cluster.