Update ready-state logic
This patch updates the ready-state logic for the Cluster and Machine
objects. A machine is not marked as ready until it is possible to
connect to the target cluster and verify the machine is represented
there as a node.

Additionally, this patch relocates the code to create the KubeConfig
secret from the Cluster's Reconcile function to the Machine's Update
function. This is because the KubeConfig secret must be created before a
Cluster can verify its ready state. If the Cluster didn't create the
KubeConfig until after it was ready, then a circular dependency would be
created, and a Cluster would never appear as ready.

Finally, several existing calls have been changed to rely on the CAPI
NodeRef/Remote packages in order to get a client for a target cluster
or retrieve a KubeConfig for the target cluster.