Cloud provider credentials as secret
This patch changes the way the cloud provider obtains its credential
information. Instead of directly via the cloud config file, a Kubernetes
Secret object is created on target clusters that is used by the cloud
provider. This is a much more secure design as credentials are no longer
transported via the GuestInfo RPC mechanism as base64-encoded plain-text
data.