Have e2e.sh script fail early upon error
This patch changes the e2e.sh script to always exit when an error is
encountered. Previously there were scenarious where errors went
unnoticed and the script still exited with code 0.

A trap is added to always cleanup the created VM if it is present.

Script now calls for bash in the shebang.