Add hack/release.sh script
Create a new hack/release.sh script that can be used for building new
container images and the clusterctl binary for both prod/GA and CI use.
This script accepts arguments to optionally push the images to the
registry and cloud storage buckets, and tags the images differently if
 they are designated as only for CI.

This script is meant to be consumed by the top level Makefile in the
future, but also can be used by developers directly.