Change e2e script/job to use new image locations
This patch updates e2e.sh and related pieces to use images located at
gcr.io/cluster-api-provider-vsphere/pr instead of the previous location.

It also eliminates the need for a temporary 'clusterctl' container image
as part of the e2e process, because that binary is now being pushed as
part of the automated builds. Therefore we just download the binary
instead of building and pushing a new image just to copy the binary out
of it.