Add -b option to hack/release.sh
The -b option allows one to specify a branch for use with the -l option.
This means that if -l is used (tag images as latest) the "latest" tag
will *only* be applied if the current commit is in/on the given branch.
Otherwise "latest" will not be applied.

This allows for CI to do a job like "hack/release.sh -l -p -b
release-0.4" that runs on every commit, and only have the "latest" tag
applied to images built on tags on the "release-0-4" branch.