Fixes error when using -P flag with generator
This patch fixes an error when using the -P flag with the generator to
set the CAPI manager's log level. The getopts string was missing a :
character after the P, and so the shell was saying OPTARG was an unbound
variable since the -P flag didn't pick up its value.