Reconcile based on task completion
This patch introduces the ability to trigger reconcile events for any
resource based on external events, such as when a task completes.

This patch removes the DefaultRequeue logic for the VSphereMachine
controller and replaces it by triggering a GenericEvent for a
VSphereMachine when an associated task for the VSphereMachine's back-end
VM completes (successfully or otherwise).