VSphereVM resource / controller
This patch introduces a new CRD, VSphereVM. This CRD is used to
provision new VMs on vSphere, and the VSphereMachine controller has been
refactored to use this method. The reason for this change is to support
deploying the HAProxy load balancer OVA using the existing GoVmomi
infrastructure in place to deploy the k/k machine images.