Build HAProxy OVA with Photon 3 Rev2 minimal
This patch uses the Photon 3 Rev2 minimal ISO to build the HAProxy OVA.
This reduces the size of the ISO downloaded to build the OVA from +5GiB
to ~300MiB.

One aspect of this patch requires installing additional packages in the
post-install script in the kickstart. While the additional_packages
element is still present, its list of packages is installed with tdnf in
the script at the bottom of the kickstart file. This is because the
Photon minimal ISO consults only itself when processing the
additional_packages directive, and any package not found on the minimal
ISO, including "sudo", will not be installed.