Fix VSphereCluster VCenterAvailable condition
We now check the connectivity regardless of whether
CloudProviderConfiguration is set on the spec.

That property has been deprecated and it is valuable to check this
connectivity early and surface this on the VSphereCluster status rather
than just on VSphereVMs.

This also fixes a bug where the status could be set to "False" if the
secret was not found and then never properly set to "True" once the secret
becomes present.

https://github.com/kubernetes-sigs/cluster-api-provider-vsphere/issues/1209

Co-authored-by: David McClure <dmcclure@vmware.com>
Co-authored-by: Tyler Schultz <tschultz@vmware.com>