Adds logic for deletion of orphaned VSphereMachine
This patch adds the logic for the deletion of orphaned vspheremachine
objects. Orphaned VSphereMachine objects are those who do not have the
CAPI Machine objects set as their owners. This situation could arise if
the owner CAPI Machine gets deleted before claiming the VSphereMachine
object as its dependent object.
There is a CAPI bug which is supposed to solve this issue, but in the
meanwhile, if the VSPhereMachine does not have any other owner
references apart from the VSphereCluster object, then we delete those
objects during VSphereCluster deletion loop.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>