Update golangci-lint to 1.43.0
This updates the version of golangcli-lint we are using to 1.43.0. This
version picks up linter releases that address issues when running under
go 1.17. This prepares us for eventually raising the required golang
version for the project and makes sure developers that have newer
versions installed can run local tests without issues.

The newer versoin of golangci-lint brings with it additional checks that
were not enforced before. Due to that, in addition to raising the
golangci-lint version this also fixes a lot of (most minor) things that
the newer version uncovered.

This touchs a very large number of files. In almost all cases though,
the changes in these files are very minor cosmetic updates.

Signed-off-by: Sean McGinnis <smcginnis@vmware.com>