Enqueue VSphereMachine in clusterToVSphereMachines mapper
The clusterToVSphereMachines mapper gets a list of machines in a cluster
and enqueues objects using the Machine's name instead of the corresponding
VSphereMachine's name. If the VSphereMachine and Machine object are named differently,
the controller can't reconcile the VSphereMachine object since it tries
to find it by the wrong name. This commit changes the mapper to enqueue
the corresponding VSphereMachine of a Machine by using InfrastructureRef.Name
field