Fixes secret fetch logic for vmware controllers
The provider service account and service discovery controllers had an
inherent assumption that the names of the Cluster and VSphereCluster
object is the same. Hence the logic to fetch the secret containing the
kubeconfig of the guest cluster was leveraging the name of the
VSphereCluster object.
With the advent of ClusterClass, the name of the generated
VSphereCluster object is different than the name of the Cluster object.

This patch adds the logic to get the owner Cluster object from the
VSphereCluster object and use that to generate the name of the secret
that contains the kubeconfig.
It also updates the test logic by passing different name to the Cluster
object.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>