Removing the ensure-go check from ci-test-coverage.sh
Currently there is a go version compatibility issue.
Test infra job is failing as go version in kubekins image is 1.17
while in dependencies like clusterapi@test is 1.16 and go mod
version in hack/tools is 1.15. This is causing gomodules to throw
error and hence failing the test coverage job.
Removing this line as kubekins by default has 1.17 go installed.
This script will run assuming that go will be present in the kubekins.

Signed-off-by: geetikab@vmware.com <geetikab@vmware.com>