Switch to CreateOrPatch; ignore manager binary
This patch updates instances of CreateOrUpdate to CreateOrPatch
in order to avoid being destructive during mutation of existing
resources.

This patch also updates the project's .gitignore file to ignore
the manager binary that gets created at the root of the project
if "go build" is used to build the manager.