Fixes reconcile request mechanism
This patch fixes the logic of enqueuing requests to the VSphereCluster
controller when VSphereMachine objects get changed. This is needed since
the name of the Cluster object is different than the name of the
VsphereCluster object in case of topology controller.

It also cleans up the unused GetMachineInCluster() function which can be
replaced by GetFilteredMachinesForCluster() function from the CAPI util
package.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>