Adds ESXi host info to VSphere Machine
This patch introduces a new field to the Status of the VSphereVM object
which contains the information of the host that the VM is placed on.
This information is propagated to the VSphereMachine object and then
added to the Machine object as a label with key
node.cluster.x-k8s.io/esxi-host.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>