Fixes panic when creating VSphereVM
This fixes a panic that was caused because of a race condition between
the VSphere machine controller creating the VM which was causing another
reconcile event being requeued for the same object. The other reconcile
event was sometimes failing to find the created VSphereVM from the
earlier request and trying to create the same object and failing with an
already exists error.
Now, the VSphereMachine controller omits the create events for the
VSphereVM objects since the controller is the author of those objects
and should not be notified for the same event.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>