Adds feature flag for node labeling
This patch introduces a new feature flag which gates the node labeling
functionality. It controls whether the control responsible for labeling
the node objects with special prefixed labels is added to the manager on
startup.
The name of the feature flag is `NodeLabeling` and can be set/reset by
setting the `EXP_NODE_LABELING` to true/false.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>