Fixes feature flag import
This is a subset of the changes in the PR #1602 that was merged in the
main branch. This specifically updates the feature gate import so that
we are not by default exposing the CAPI feature flags on CAPV which are
not usable.
This is needed since node anti affinity and node labeling
functionalities are under separate and 2 new feature flags

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>