Adds default resource pool logic
This patch adds the logic to fall back to the default resource pool, if
no resource pool is specified, during cluster module creation.
It also checks whether the resource pool is owned by a compute cluster
or a standalone host. In case of the owner being a standalone host, a
warning is introduced on the VSphereCluster object and the cluster
creation proceeds.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>