Adds provision to set VMX Version
Currently, during cloen operations the VMX version set in the OVA is
used by the created VM.
This patch provides a way to input the VMX version that can be set on
the cloned VM object. The patch assumes that the user has checked the
ESXi host compatibility with the supplied version.

Signed-off-by: geetikab@vmware.com <geetikab@vmware.com>
Signed-off-by: Sagar Muchhal <muchhals@vmware.com>