Checks for feature flag during VM reconciliation
Adds the check for the state of the anti affinity feature flag when
reconciling VM. When reconciling the VM, the cluster module info should
be looked up only when the NodeAntiAffinity feature flag is turned on.
This also solves the issue when the control plane provider is something
other than the KubeadmControlPlane, in which case we do not want to
allow checks for cluster module info.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>