Reorders the PCI device attachment process
This patch reorders the process of attaching PCI devices to the VM.
Previously, the PCI devices were attached during the VM cloning
operation, but the PCI device attachment requires a minimum hardware
version of vmx-17.
Instead of forcing the template to be hardcoded with a minimum hardware
version, CAPV introduced the capability of upgrading the hardware
version of the VM post clone operation succeeds. This would mean that
the PCI devices need to be attached after the upgrade hardware VM task
is finished and the hardware version of the VM has been successfully
upgraded to some compatible version specified by the user.