Drops version check when tracking owners
During the VM reconciliation, the controllers track down the parent of
the Machine objects using the owner reference hierarchy. The owner
reference checks were comparing the Group as well as the API version to
determins and fetch the owner.

Currently, the API version used for comparison was selected from the
storageVersion of the CAPI objects on the API server. Due to this, any
objects created via an earlier CAPI version, say v1alpha3, which sets
the owner reference API version to group name + v1alpha3 were not being
found since no such object would exist. A similar object with the
updated API type is stored in the api server.

This drify happens since CAPI does not update the owner references of
the objects when moving to newer API versions. There is an issue
tracking this https://github.com/kubernetes-sigs/cluster-api/issues/7224
which will solve the stale owner ref API version problem. Until then we
are dropping the version check and just relying on the group info and
kind combination to fetch the parent object.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>