Fix vCenter session leaks
On controller startup, the reconciliation loops for the VSphereCluster
and VSphereVM objects make a call to the seesion.GetOrCreate function
to look for existing sessions. Without the mutex guarding the access to
the function, all the objects end up creating a session, and the last
created session which gets saved in the cache gets used for future API
calls and the other sessions are left to idle.
Also, altered the logout logic to destroy the session using the govmomi
method calls directly.

Another improvement made here was to destroy the sessions on managaer
shutdown so that CAPV does not leave over any sessions on manager exit.