Updates disk size for clone test
The PR https://github.com/vmware/govmomi/pull/2911 introduces a
validation which does not allow disk resize operations to lower the size
of the disk. After the govmomi dependency update, the current setup of
the clone test setup was resizing the disk to a smaller value. Hence the
test setup had to be updated to update the disk size from 10GB to 20GB.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>