fix: Calculate default primary disk size from base template
The primary disk size is optional, according to the API.

Prior to this commit, if the user did not set an explicit primary disk
size, VM creation would fail with a validation error.

This commit sets the primary disk size to the base template primary disk
size when the user does not set an explicit primary disk size.