Marking KCP remediation test as Pending
The thought is to skip this test since this test requires a sepcialized
setup and using the CAPI defaults which work for CAPD might not be
suitable for CAPV since provisioning takes slightly longer in vSphere.

This test will need to be reworked to create a HA control plane and then
mark a node for remediation.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>