prevent IPAM addresses from being assigned to multiple devices
we had previously made an assumption that the
VSphereVM.Spec.Network.Devices[*].MACAddr would be set. This caused us
to generate IPAMState that could be keyed on an empty string and later
read for every device without a MACAddr set.

we now rely on the NetworkStatus as the source of truth for MACAddr to
key and later retrieve the correct IPAM configuration for devices.

Signed-off-by: Aidan Obley <aobley@vmware.com>
Co-authored-by: Edwin Xie <exie@vmware.com>
Co-authored-by: Tyler Schultz <tschultz@vmware.com>
Co-authored-by: Christian Ang <angc@vmware.com>