Introduces parameter for CP zone selection
This patch introduces a new parameter for selecting zones for control
plane nodes.
In the past, all zones having the same vCenter URL in the
spec.server would be associated to a vSphereCluster object with the same
vCenter URL set in the spec.server.
Now the vSphereCluster spec exposes a label based selection which can be
used to choose the zones for control plane node placement. The label
selector follows the same rules as described on the label selector types in
https://pkg.go.dev/k8s.io/apimachinery@v0.26.1/pkg/apis/meta/v1#LabelSelector

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>