Removes the v1alpha2 VMOp CRDs
VMOp recently introduced the v1a2 CRDs and flipped the default storage
version flag on the new CRDs to true and set the storage version on the
v1a1 CRDs as false. Bringing in the latest version of VMOp API, causes
the generated VMOp CRDs in CAPV to have the same change.
The VMOp APIs between v1a1 and v1a2 have changed considerably, and it is
the responsibility of VMOp's conversion webhooks to make sure that the
functionality remains the same even when the storage version for v1a2 is
flipped to true.

Currently CAPV explcitly uses the VMOp v1a1 CRDs as the base for
creating the custom resources. As part of the integration tests for the
supervisor API types, CAPV runs a set of integration tests on the ind
cluster. No conversion webhooks are operational in this test setup.
Hence, we cannot rely on the auto conversion of CRs across versions via
the VMOp webhooks. For this reason, we are explictly generating the
v1alpha1 CRDs only and using those in the integration tests.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>