[release-1.12] 🐛 fix: use machine's .spec.failureDomain with fallback to vspheremachine's .spec.failureDomain if not defined
This commit sets failureDomain string defaulting to
`machine.Spec.FailureDomain`. If failureDomain
is an empty string then `vsphereMachine.Spec.FailureDomain` is
checked for nil, if defined `failureDomain` is set
to `vsphereMachine.Spec.FailureDomain`.

This resolves an issue where virtual machines are not placed
within vm-host groups in a timely
manner causing misplacement in the topology.