Remove invalid kustomizeconfig
This change has no effect on the output of this kustomization because
the removed configuration was redundant. However, it fixes a bug which
can be triggered when using this kustomization as a base for another
kustomization.

We remove 2 types of configuration. The first is redundant, the second
is incorrect and triggers the bug.

Firstly we remove nameReference configurations which specify references
which are already part of kustomize's defaults. These were found in:
* config/govmomi/webhook/kustomizeconfig.yaml
* config/supervisor/webhook/kustomizeconfig.yaml

Secondly we remove all namespace configurations as these were all
incorrect. These directed the namespace transformer to update
webhooks/clientConfig/service/namespace. However, this is not the
intended function of the namespace transformer: it should only set the
namespace directly on objects and allow references to be updated
automatically by nameReference. Configuring it to update a reference
directly leaves kustomize with inconsistent internal state. Depending on
execution order this can cause a subsequent transformation to fail to
update the reference when it makes further changes to the Service
object.